+++ src/repetition_info.cpp
@@ -1,5 +1,5 @@
 #include "repetition_info.hpp"
-#include <cstddef>
+#include "util/types.hpp"
 
 namespace Clockwork {
 
@@ -11,7 +11,7 @@ void RepetitionInfo::pop() {
     m_repetition_table.pop_back();
 }
 
-bool RepetitionInfo::detect_repetition(size_t root_ply) {
+bool RepetitionInfo::detect_repetition(usize root_ply) {
     // Short circuit return if move is not reversible
     if (!m_repetition_table.back().second) {
         return false;
@@ -19,18 +19,18 @@ bool RepetitionInfo::detect_repetition(size_t root_ply) {
 
     // The hash we will be looking for
     HashKey query  = m_repetition_table.back().first;
-    size_t  height = m_repetition_table.size() - 1;
+    usize   height = m_repetition_table.size() - 1;
 
     // Count how many times we encounter the hash
     i32 counter = 0;
 
     // Start from two moves ago
-    for (size_t idx = 4; idx <= height; idx += 2) {
+    for (usize idx = 4; idx <= height; idx += 2) {
         // Do the hashes match?
         if (m_repetition_table[height - idx].first == query) {
 
             // If the match happened inside the search tree, twofold repetition check
-            if (idx <= root_ply) {
+            if (idx <= static_cast<usize>(root_ply)) {
                 return true;
             }
